third_party/WebKit/LayoutTests/fast/css/quotes-computedstyle.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>

<style>
q:lang(no) { quotes: none }
q:lang(qs) { quotes: "<<" ">>" "<" ">" }
</style>

<p id="p1">Before <q lang="no">q1-text <q>q2-text</q> q1-text</q> after.</p>
<p id="p2">Before <q lang="qs">q1-text <q>q2-text</q> q1-text</q> after.</p>

<script>

description("Tests the computed style of quotes property.");
var child1 = document.getElementById('p1').children[0].children[0];
var child2 = document.getElementById('p2').children[0].children[0];

shouldBeEqualToString("getComputedStyle(child1).quotes","none");
shouldBeEqualToString("getComputedStyle(child1).getPropertyValue('quotes')", "none");

shouldBeEqualToString("getComputedStyle(child2).quotes",'"<<" ">>" "<" ">"');
shouldBeEqualToString("getComputedStyle(child2).getPropertyValue('quotes')", '"<<" ">>" "<" ">"');
</script>
