third_party/blink/web_tests/fast/css/quotes-computedstyle-expected.txt

Tests the computed style of quotes property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(child1).quotes is "none"
PASS getComputedStyle(child1).getPropertyValue('quotes') is "none"
PASS getComputedStyle(child2).quotes is "\"<<\" \">>\" \"<\" \">\""
PASS getComputedStyle(child2).getPropertyValue('quotes') is "\"<<\" \">>\" \"<\" \">\""
PASS successfullyParsed is true

TEST COMPLETE


Before q1-text q2-text q1-text after.

Before q1-text q2-text q1-text after.
